# booking-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking Billiard API",
        "description": "Welcome to the official **Booking Billiard API** documentation.\n\nThis API provides an enterprise-grade backend infrastructure designed to streamline and automate billiard center operations. Built with a focus on high performance, security, and scalability, it serves as the core engine for modern booking applications.\n\n### Core Capabilities:\n- **Secure Authentication**: Robust role-based access control (RBAC) powered by Laravel Sanctum, ensuring data privacy and operational security.\n- **Resource Management**: Dynamic and highly configurable table allocation, capacity planning, and scheduling algorithms.\n- **Advanced Reservation System**: Sophisticated booking engine capable of handling future reservations, real-time availability conflict resolution, and instant walk-in assignments.\n- **Transaction Architecture**: Seamless and transactional payment state management, ensuring financial data integrity.\n- **Analytics & Telemetry**: Comprehensive real-time monitoring of table states and data-driven reporting for business intelligence.\n\n---\n\n*Architected and developed by **Wisnu Nugraha**.*",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/tables/{table}/schedules": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Get schedules for a table",
                "operationId": "32f0dbcf6f0fa7f3d9eee6478bb20619",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ScheduleResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Add a new schedule to a table (Admin)",
                "operationId": "4bcf60d4bc0c1446e17c3a51a71d6943",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "day_of_week",
                                    "open_time",
                                    "close_time"
                                ],
                                "properties": {
                                    "day_of_week": {
                                        "type": "integer"
                                    },
                                    "open_time": {
                                        "type": "string",
                                        "format": "time"
                                    },
                                    "close_time": {
                                        "type": "string",
                                        "format": "time"
                                    },
                                    "is_available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Jadwal berhasil ditambahkan."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ScheduleResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error or duplicate schedule"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tables/{table}/schedules/{schedule}": {
            "put": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Update a schedule (Admin)",
                "operationId": "3c0d12dac3afdc5bc89075b9bc3e1b08",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "schedule",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "day_of_week": {
                                        "type": "integer"
                                    },
                                    "open_time": {
                                        "type": "string",
                                        "format": "time"
                                    },
                                    "close_time": {
                                        "type": "string",
                                        "format": "time"
                                    },
                                    "is_available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Jadwal berhasil diperbarui."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ScheduleResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Delete a schedule (Admin)",
                "operationId": "15706767eddaeedfadb6ac4560a535a2",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "schedule",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Jadwal berhasil dihapus."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tables/{table}/availability": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Check table availability for a specific date",
                "operationId": "aff849cf23d0e51f798526d7bc48422b",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "date": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "day_name": {
                                                    "type": "string"
                                                },
                                                "is_available": {
                                                    "type": "boolean"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "open_time": {
                                                    "type": "string",
                                                    "format": "time",
                                                    "nullable": true
                                                },
                                                "close_time": {
                                                    "type": "string",
                                                    "format": "time",
                                                    "nullable": true
                                                },
                                                "slots": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "start_time": {
                                                                "type": "string",
                                                                "format": "time"
                                                            },
                                                            "end_time": {
                                                                "type": "string",
                                                                "format": "time"
                                                            },
                                                            "is_available": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ScheduleResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "table_id": {
                        "type": "integer"
                    },
                    "day_of_week": {
                        "type": "integer"
                    },
                    "day_name": {
                        "type": "string"
                    },
                    "open_time": {
                        "type": "string",
                        "format": "time"
                    },
                    "close_time": {
                        "type": "string",
                        "format": "time"
                    },
                    "is_available": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}